I #1248: Updated stub generation tests for scalars
